# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Clinic Management System API",
    "description": "Comprehensive API for managing multi-clinic healthcare operations",
    "version": "1.0.0",
    "contact": {
      "email": "support@clinicmanagement.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    },
    {
      "url": "https://localhost:5000/"
    }
  ],
...
  "paths": {
...
    "/api/doctor/payments": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinic_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/doctor/appointments/all": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinic_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/doctor/appointments/details/{appointment_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/doctor/appointments/update/{appointment_id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "example": "any"
                  },
                  "notes": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/doctor/appointments/patient/{patient_id}/records": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctor_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinic_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/doctor/appointments/dashboard": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinic_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/doctor/appointments/today": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinic_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/doctor/appointments/debug": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinic_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/doctor/appointments/payments": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "doctor_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinic_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}